test(openai): 按模块拆分 OpenAI 测试
OpenAI 测试 suite 和辅助类型此前集中在 ServiceTests.swift 中，难以按模块定位，也不符合新的测试目录约定。
本阶段只调整测试组织和工程登记，不改变产品行为。

将 4 个 OpenAI suite 拆分到 EasydictTests/Service/OpenAI/，保留全部测试语义、tags、流式 transport 的 serialized 边界和局部 helpers。
同步更新 Xcode 测试分组、文件引用和 target Sources，并补充测试目录规则。

拆分后的 4 个 suite 共 11 个测试继续通过，测试名称和工程引用与拆分前保持一致。
同时归档本次执行计划并完成 history，形成完整的移植记录。

----------------------------------------------------------------------

test(openai): split OpenAI tests by module

The OpenAI test suites and helpers were previously concentrated in ServiceTests.swift, making them harder to locate by module and inconsistent with the new test directory convention.
This stage changes only test organization and project registration, without changing product behavior.

Split the four OpenAI suites into EasydictTests/Service/OpenAI/ while preserving all test semantics, tags, the serialized streaming transport boundary, and local helpers.
Update the Xcode test group, file references, and target Sources, and document the test directory rules.

The four split suites continue to pass all 11 tests, with test names and project references matching the pre-split layout.
Archive the execution plan and complete the history to retain a full migration record.
